+++ rosu-v2/src/model/matches.rs
@@ -1,5 +1,6 @@
 use super::{
-    beatmap::Beatmap, mods::GameMods, score_::LegacyScoreStatistics, serde_, user_::User, GameMode,
+    beatmap::Beatmap, mods::GameMods, score::LegacyScoreStatistics, serde_util, user::User,
+    GameMode,
 };
 use crate::{
     prelude::{GameModsIntermode, ModeAsSeed},
@@ -26,7 +27,7 @@ pub enum MatchEvent {
     Create {
         #[cfg_attr(feature = "serialize", serde(rename(serialize = "id")))]
         event_id: u64,
-        #[cfg_attr(feature = "serialize", serde(with = "serde_::datetime"))]
+        #[cfg_attr(feature = "serialize", serde(with = "serde_util::datetime"))]
         timestamp: OffsetDateTime,
         user_id: Option<u32>,
     },
@@ -35,7 +36,7 @@ pub enum MatchEvent {
     Disbanded {
         #[cfg_attr(feature = "serialize", serde(rename(serialize = "id")))]
         event_id: u64,
-        #[cfg_attr(feature = "serialize", serde(with = "serde_::datetime"))]
+        #[cfg_attr(feature = "serialize", serde(with = "serde_util::datetime"))]
         timestamp: OffsetDateTime,
     },
     /// A map is / was being played
@@ -48,7 +49,7 @@ pub enum MatchEvent {
         game: Box<MatchGame>,
         #[cfg_attr(feature = "serialize", serde(default))]
         match_name: String,
-        #[cfg_attr(feature = "serialize", serde(with = "serde_::datetime"))]
+        #[cfg_attr(feature = "serialize", serde(with = "serde_util::datetime"))]
         timestamp: OffsetDateTime,
     },
     /// The match host changed
@@ -56,7 +57,7 @@ pub enum MatchEvent {
     HostChanged {
         #[cfg_attr(feature = "serialize", serde(rename(serialize = "id")))]
         event_id: u64,
-        #[cfg_attr(feature = "serialize", serde(with = "serde_::datetime"))]
+        #[cfg_attr(feature = "serialize", serde(with = "serde_util::datetime"))]
         timestamp: OffsetDateTime,
         user_id: u32,
     },
@@ -65,7 +66,7 @@ pub enum MatchEvent {
     Joined {
         #[cfg_attr(feature = "serialize", serde(rename(serialize = "id")))]
         event_id: u64,
-        #[cfg_attr(feature = "serialize", serde(with = "serde_::datetime"))]
+        #[cfg_attr(feature = "serialize", serde(with = "serde_util::datetime"))]
         timestamp: OffsetDateTime,
         user_id: u32,
     },
@@ -74,7 +75,7 @@ pub enum MatchEvent {
     Kicked {
         #[cfg_attr(feature = "serialize", serde(rename(serialize = "id")))]
         event_id: u64,
-        #[cfg_attr(feature = "serialize", serde(with = "serde_::datetime"))]
+        #[cfg_attr(feature = "serialize", serde(with = "serde_util::datetime"))]
         timestamp: OffsetDateTime,
         user_id: u32,
     },
@@ -83,7 +84,7 @@ pub enum MatchEvent {
     Left {
         #[cfg_attr(feature = "serialize", serde(rename(serialize = "id")))]
         event_id: u64,
-        #[cfg_attr(feature = "serialize", serde(with = "serde_::datetime"))]
+        #[cfg_attr(feature = "serialize", serde(with = "serde_util::datetime"))]
         timestamp: OffsetDateTime,
         user_id: u32,
     },
@@ -208,7 +209,7 @@ impl<'de> Visitor<'de> for MatchEventVisitor {
 
     fn visit_map<A: MapAccess<'de>>(self, mut map: A) -> Result<Self::Value, A::Error> {
         #[derive(Deserialize)]
-        struct DateTimeWrapper(#[serde(with = "serde_::datetime")] OffsetDateTime);
+        struct DateTimeWrapper(#[serde(with = "serde_util::datetime")] OffsetDateTime);
 
         let mut id = None;
         let mut timestamp: Option<DateTimeWrapper> = None;
@@ -299,13 +300,13 @@ impl<'de> Deserialize<'de> for MatchEvent {
 pub struct MatchGame {
     #[cfg_attr(feature = "serialize", serde(rename = "id"))]
     pub game_id: u64,
-    #[cfg_attr(feature = "serialize", serde(with = "serde_::datetime"))]
+    #[cfg_attr(feature = "serialize", serde(with = "serde_util::datetime"))]
     pub start_time: OffsetDateTime,
     #[cfg_attr(
         feature = "serialize",
         serde(
             skip_serializing_if = "Option::is_none",
-            with = "serde_::option_datetime"
+            with = "serde_util::option_datetime"
         )
     )]
     pub end_time: Option<OffsetDateTime>,
@@ -326,9 +327,9 @@ impl<'de> Deserialize<'de> for MatchGame {
         struct MatchGameRawMods {
             #[serde(rename = "id")]
             game_id: u64,
-            #[serde(with = "serde_::datetime")]
+            #[serde(with = "serde_util::datetime")]
             start_time: OffsetDateTime,
-            #[serde(with = "serde_::option_datetime")]
+            #[serde(with = "serde_util::option_datetime")]
             end_time: Option<OffsetDateTime>,
             mode: GameMode,
             scoring_type: ScoringType,
@@ -463,13 +464,13 @@ pub struct MatchInfo {
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
-        with = "serde_::option_datetime"
+        with = "serde_util::option_datetime"
     )]
     pub end_time: Option<OffsetDateTime>,
     #[serde(rename = "id")]
     pub match_id: u32,
     pub name: String,
-    #[serde(with = "serde_::datetime")]
+    #[serde(with = "serde_util::datetime")]
     pub start_time: OffsetDateTime,
 }
 
@@ -523,7 +524,7 @@ pub struct MatchListParams {
 #[cfg_attr(feature = "serialize", derive(serde::Serialize))]
 pub struct MatchScore {
     /// Accuracy between `0.0` and `100.0`
-    #[cfg_attr(feature = "serialize", serde(with = "serde_::adjust_acc"))]
+    #[cfg_attr(feature = "serialize", serde(with = "serde_util::adjust_acc"))]
     pub accuracy: f32,
     pub max_combo: u32,
     pub mods: GameModsIntermode,
@@ -664,7 +665,7 @@ pub struct OsuMatch {
         feature = "serialize",
         serde(
             skip_serializing_if = "Option::is_none",
-            with = "serde_::option_datetime"
+            with = "serde_util::option_datetime"
         )
     )]
     pub end_time: Option<OffsetDateTime>,
@@ -673,7 +674,7 @@ pub struct OsuMatch {
     pub latest_event_id: u64,
     pub match_id: u32,
     pub name: String,
-    #[cfg_attr(feature = "serialize", serde(with = "serde_::datetime"))]
+    #[cfg_attr(feature = "serialize", serde(with = "serde_util::datetime"))]
     pub start_time: OffsetDateTime,
     /// Maps user ids to users
     #[cfg_attr(feature = "serialize", serde(serialize_with = "serialize_match_users"))]
@@ -877,11 +878,11 @@ impl<'de> Visitor<'de> for OsuMatchVisitor {
 
     fn visit_map<A: MapAccess<'de>>(self, mut map: A) -> Result<Self::Value, A::Error> {
         #[derive(Deserialize)]
-        struct DateTimeWrapper(#[serde(with = "serde_::datetime")] OffsetDateTime);
+        struct DateTimeWrapper(#[serde(with = "serde_util::datetime")] OffsetDateTime);
 
         #[derive(Deserialize)]
         struct OptionDateTimeWrapper(
-            #[serde(with = "serde_::option_datetime")] Option<OffsetDateTime>,
+            #[serde(with = "serde_util::option_datetime")] Option<OffsetDateTime>,
         );
 
         let mut current_game_id = None;
