+++ rosu-v2/src/model/score.rs
@@ -1,7 +1,7 @@
 use super::{
-    beatmap_::{BeatmapExtended, Beatmapset},
+    beatmap::{BeatmapExtended, Beatmapset},
     mods::GameMods,
-    serde_,
+    serde_util,
     user::User,
     GameMode, Grade,
 };
@@ -58,10 +58,10 @@ pub struct Score {
     #[cfg_attr(feature = "serialize", serde(rename = "type"))]
     pub kind: Box<str>,
     pub user_id: u32,
-    #[cfg_attr(feature = "serialize", serde(with = "serde_::adjust_acc"))]
+    #[cfg_attr(feature = "serialize", serde(with = "serde_util::adjust_acc"))]
     pub accuracy: f32,
     pub build_id: Option<u32>,
-    #[cfg_attr(feature = "serialize", serde(with = "serde_::datetime"))]
+    #[cfg_attr(feature = "serialize", serde(with = "serde_util::datetime"))]
     pub ended_at: OffsetDateTime,
     pub has_replay: bool,
     pub is_perfect_combo: bool,
@@ -74,7 +74,7 @@ pub struct Score {
     pub pp: Option<f32>,
     #[cfg_attr(feature = "serialize", serde(rename = "ruleset_id"))]
     pub mode: GameMode,
-    #[cfg_attr(feature = "serialize", serde(with = "serde_::option_datetime"))]
+    #[cfg_attr(feature = "serialize", serde(with = "serde_util::option_datetime"))]
     pub started_at: Option<OffsetDateTime>,
     #[cfg_attr(feature = "serialize", serde(rename = "total_score"))]
     pub score: u32,
@@ -112,10 +112,10 @@ impl<'de> Deserialize<'de> for Score {
             #[serde(rename = "type")]
             kind: Box<str>,
             user_id: u32,
-            #[serde(with = "serde_::adjust_acc")]
+            #[serde(with = "serde_util::adjust_acc")]
             accuracy: f32,
             build_id: Option<u32>,
-            #[serde(alias = "created_at", with = "serde_::datetime")]
+            #[serde(alias = "created_at", with = "serde_util::datetime")]
             ended_at: OffsetDateTime,
             has_replay: Option<bool>,       // not available in legacy scores
             is_perfect_combo: Option<bool>, // not available in legacy scores
@@ -131,7 +131,7 @@ impl<'de> Deserialize<'de> for Score {
             _mode: Option<IgnoredAny>, // only available in legacy scores
             #[serde(rename = "ruleset_id", alias = "mode_int")]
             mode: GameMode,
-            #[serde(default, with = "serde_::option_datetime")]
+            #[serde(default, with = "serde_util::option_datetime")]
             started_at: Option<OffsetDateTime>,
             #[serde(rename = "total_score", alias = "score")]
             score: u32,
@@ -299,37 +299,37 @@ pub struct ScoreStatistics {
     #[serde(
         default,
         alias = "count_miss",
-        deserialize_with = "serde_::from_option::deserialize"
+        deserialize_with = "serde_util::from_option::deserialize"
     )]
     pub miss: u32,
     #[serde(
         default,
         alias = "count_50",
-        deserialize_with = "serde_::from_option::deserialize"
+        deserialize_with = "serde_util::from_option::deserialize"
     )]
     pub meh: u32,
     #[serde(
         default,
         alias = "count_100",
-        deserialize_with = "serde_::from_option::deserialize"
+        deserialize_with = "serde_util::from_option::deserialize"
     )]
     pub ok: u32,
     #[serde(
         default,
         alias = "count_katu",
-        deserialize_with = "serde_::from_option::deserialize"
+        deserialize_with = "serde_util::from_option::deserialize"
     )]
     pub good: u32,
     #[serde(
         default,
         alias = "count_300",
-        deserialize_with = "serde_::from_option::deserialize"
+        deserialize_with = "serde_util::from_option::deserialize"
     )]
     pub great: u32,
     #[serde(
         default,
         alias = "count_geki",
-        deserialize_with = "serde_::from_option::deserialize"
+        deserialize_with = "serde_util::from_option::deserialize"
     )]
     pub perfect: u32,
     #[serde(default)]
